requirements.txt
mysql-connector-python
aiogram3
requests
docxtpl
datetime